requirements.txt
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
pydantic==2.1.1
python-dotenv==1.0.0
social-auth-app-django==5.2.0
